feat: 음표 모델 코드 수정
